fix: remove nested computed field from model
